requirements.txt
accelerate==1.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
coloredlogs==15.0.1
filelock==3.17.0
flatbuffers==25.2.10
fsspec==2025.2.0
huggingface-hub==0.28.1
humanfriendly==10.0
idna==3.10
jinja2==3.1.5
markupsafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.3
onnxruntime==1.20.1
packaging==24.2
protobuf==5.29.3
psutil==7.0.0
pycparser==2.22
pyopenjtalk-plus==0.3.4.post10
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
sentencepiece==0.2.0
soundfile==0.13.1
sudachidict-core==20250129
sudachipy==0.6.10
sympy==1.13.1
tokenizers==0.13.3
torch==2.6.0
torchaudio==2.6.0
tqdm==4.67.1
transformers==4.29.2
typing-extensions==4.12.2
urllib3==2.3.0
